+++ cinema-adapter/src/main/java/com/cinema/adapter/in/web/RateLimitingInterceptor.java
@@ -1,25 +1,27 @@
-package com.cinema.adapter.in.web.filter;
+package com.cinema.adapter.in.web;
 
 import com.cinema.application.port.out.RateLimiterPort;
 import com.cinema.domain.exception.CoreException;
 import com.cinema.domain.exception.ErrorType;
-import jakarta.servlet.*;
+import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Component;
+import org.springframework.web.servlet.HandlerInterceptor;
 
 import java.io.IOException;
 
 @Component
 @RequiredArgsConstructor
-public class RateLimitingFilter implements Filter {
+public class RateLimitingInterceptor implements HandlerInterceptor {
 
     private final RateLimiterPort rateLimiterPort;
 
     @Override
-    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
-        HttpServletRequest httpRequest = (HttpServletRequest) servletRequest;
-        String ip = httpRequest.getRemoteAddr();
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+        String ip = request.getRemoteAddr();
 
         if (rateLimiterPort.isBlocked(ip)) {
             throw new CoreException(ErrorType.IP_BLOCKED, "요청 IP: " + ip);
@@ -30,6 +32,6 @@ public class RateLimitingFilter implements Filter {
             throw new CoreException(ErrorType.IP_BLOCKED, "요청 IP: " + ip);
         }
 
-        filterChain.doFilter(servletRequest, servletResponse);
+        return true;
     }
 }
