+++ cinema-adapter/src/main/java/com/cinema/adapter/in/web/config/WebMvcConfig.java
@@ -1,0 +1,20 @@
+package com.cinema.adapter.in.web.config;
+
+import com.cinema.adapter.in.web.RateLimitingInterceptor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@RequiredArgsConstructor
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    private final RateLimitingInterceptor rateLimitingInterceptor;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(rateLimitingInterceptor)
+                .addPathPatterns("/api/v1/movies/**");
+    }
+}
