--- cinema-adapter/src/main/java/com/cinema/adapter/in/web/filter/FilterConfig.java
@@ -1,15 +1,0 @@
-package com.cinema.adapter.in.web.filter;
-
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class FilterConfig {
-
-    public FilterRegistrationBean<RateLimitingFilter> ipRateLimitingFilter(RateLimitingFilter filter) {
-        FilterRegistrationBean<RateLimitingFilter> registration = new FilterRegistrationBean<>();
-        registration.setFilter(filter);
-        registration.addUrlPatterns("/api/v1/movies/*");
-        return registration;
-    }
-}
